Dying Animation and Interactable Object UI
From now on cat gets damage and dies. I have added cute angel animation for the cat. Further, we have an UI element for interactable objects like socket. When mouse is over the interactable object it will pop-up a UI element. Also, now we know if cat is grounded or not.